+++ src/search_providers/l337x_search.rs
@@ -44,6 +44,7 @@ fn parse_l337x(mut results: Vec<TorrentSearchResult>) -> Vec<Torrent> {
                 .leeches
                 .map_err(|e| warn!("Getting leechers failed: {}", e))
                 .ok(),
+            ..Default::default()
         });
     }